Inline the readonly struct SqlTypeInfo.
It's only ever a temporary subexpression anyhow
